+++ render/src/cx_dx11.rs
@@ -287,7 +287,7 @@ impl Cx {
         }
     }
     
-    pub fn send_signal(signal: Signal, value: u64) {
+    pub fn send_signal(signal: Signal, value: usize) {
         Win32App::post_signal(signal.signal_id, value);
     }
 }
